% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add_points}
\alias{add_points}
\title{Add points}
\usage{
add_points(xy, landscape, width)
}
\arguments{
\item{xy}{Coordinates of the points.}

\item{landscape}{Polygon of the focal landscape.}

\item{width}{Parameter for the IIC calculation.}
}
\value{
A new polygon, similiar to the inputed, but with an small point
with additional habitat.
}
\description{
This function implements an algorithm that add a small habitat portion to a
given landscape.
}
\examples{
pts = rbind(c(-45,23))
l = rgdal::readOGR('./path to shapefile directory/')
add_points(pts, l, 0.1)

}
