% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bottom5.R
\name{Bottom5}
\alias{Bottom5}
\title{Bottom 5 Country Function}
\usage{
Bottom5(a)
}
\arguments{
\item{a}{This should be a one of column which we want to know top 5.}
}
\value{
Returns a dataframe which includes bottom 5 countries in OECD and information which we want to know (a)
}
\description{
Bottom 5 Country Function
}
\examples{
Bottom5(OECD$AREA)
}
