% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{browse_metadata}
\alias{browse_metadata}
\title{Browse the metadata related to a series.}
\usage{
browse_metadata(dataset, ...)
}
\arguments{
\item{dataset}{A string specifying the code of the series.}

\item{...}{Additional parameters passed to browseURL.}
}
\value{
Opens a web page in the default web browser.
}
\description{
Opens up a web browser with the metadata related to the requested series.
}
\examples{
\dontrun{browse_metadata("DUR_D")}
}

