% Generated by roxygen2 (4.0.2): do not edit by hand
\name{browseMetadata}
\alias{browseMetadata}
\title{Browse the metadata related to a series.}
\usage{
browseMetadata(series, data = getDatasets())
}
\arguments{
\item{series}{A string specifying the code of the series.}

\item{data}{A data frame with the data set codes, titles, and metadata.
This can be either a data frame previously fetched using
\code{\link{getDatasets}} (recommended) or left blank, in which case a
temporary data frame is fetched. The second option adds a few seconds to
each browse request.}
}
\value{
Opens a temporary XML file in the default web browser.
}
\description{
Opens up a web browser with the metadata related to the requested series.
}
\examples{
datasets <- getDatasets()
browseMetadata("DUR_D", datasets)
}

