% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_solution.R
\name{eval_solution}
\alias{eval_solution}
\title{Evaluate solution}
\usage{
eval_solution(solution_polygon, folder_path, max_dist, habitat, metric = "IIC")
}
\arguments{
\item{solution_polygon}{The polygon for the focal landscape.}

\item{folder_path}{Path to the project directory.}

\item{max_dist}{Parameter for the IIC calculation.}

\item{habitat}{Value in which habitat is assigned in the solution_polygon.}

\item{metric}{The desired metric (currently, only implemented for IIC).}
}
\value{
The IIC metric for the imputed landscape polygon.
}
\description{
Compute the IIC (Integral Index of Connectivity) for a given landscape.
}
\examples{
l = rgdal::readOGR('./path to shapefile directory/')
eval_solution(l, './path to my project directory/', 100, 1, 'IIC')

}
