% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fossil.portion.R
\name{Fossil.portion}
\alias{Fossil.portion}
\title{The portion of generation energy using fossil fuels of OECD Countries}
\usage{
Fossil.portion(x)
}
\arguments{
\item{x}{"Country name"}
}
\value{
The percentage of using fossil fuels in the total energy generation
}
\description{
The portion of generation energy using fossil fuels of OECD Countries
}
\examples{
Fossil.portion("Canada")
}
