% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_structure}
\alias{get_data_structure}
\title{Get the data structure of a dataset.}
\usage{
get_data_structure(dataset)
}
\arguments{
\item{dataset}{A string containing the code for a dataset}
}
\value{
A list of data frames.
}
\description{
Returns a list of data frames containing names and descriptions of the variables 
of a specified series.
}
\examples{
\dontrun{get_data_structure("DUR_D")}

}
