% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{get_datasets}
\alias{get_datasets}
\title{Get a data frame with information on all available datasets.}
\usage{
get_datasets()
}
\value{
A data frame.
}
\description{
Returns a data frame with two variables: \code{id} and \code{description}
}
\examples{
#datasets <- get_datasets()
#head(datasets)
}
\seealso{
\code{\link{search_dataset}} to search for a specific data set or a
keyword in the description, and \code{\link{get_data_structure}} to get the
dimensions of specified data set.
}

