% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get the dimensions of a dataset.}
\usage{
get_dimensions(dataset)
}
\arguments{
\item{dataset}{A string containing the code for a dataset.}
}
\value{
A list of characters.
}
\description{
Returns a list of characters containing names of the dimensions
of a specified series.
}
\examples{
\dontrun{get_dimensions("DUR_D")}

}
