% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDatasets}
\alias{getDatasets}
\title{Get a data frame with information on all available datasets.}
\usage{
getDatasets()
}
\value{
A data frame.
}
\description{
Returns a data frame with three variables: DatasetCode,
DatasetTitle, and DatasetMetadata.
}
\examples{
datasets <- getDatasets()
datasets[1:10, 1:2]
}
\seealso{
\code{\link{searchSeries}} to search for a specific DatasetCode or a
keyword in the DatasetTitle, and \code{\link{getDimensions}} to get the
dimensions of specified DatasetCode.
}

