% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDesc}
\alias{getDesc}
\title{Get a list of data frames with variable descriptions.}
\usage{
getDesc(series)
}
\description{
Returns a list of data frames containing descriptions of the variable values
in the requested series.
}
\details{
The data frames returned with \code{\link{getSeries}} often contain variables
whose values are not obvious. For example,
\code{\link{getSeries("MIG_UNEMP_GENDER")}}
}

