% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDimensions}
\alias{getDimensions}
\title{Get the dimensions of a dataset.}
\usage{
getDimensions(series)
}
\arguments{
\item{series}{A string containing the code for a series.}
}
\value{
A data frame.
}
\description{
Rreturs a data frame containing names and descriptions
of the variables of a specified series.
}
\examples{
getDimensions("DUR_D")
}
\seealso{
\code{\link{getDatasets}}, \code{\link{searchSeries}}
}

