% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSeries}
\alias{getSeries}
\title{Download OECD series.}
\usage{
getSeries(series, country = "all", since = NULL)
}
\arguments{
\item{series}{A string with the code for the desired series.}

\item{country}{A character vector with iso3c codes
\code{(e.g. c("USA", "DEU"))}.}

\item{since}{First year of data.}
}
\value{
Data frame with country-year observations.
}
\description{
Returns a data frame with the requested data, downloaded from the OECD's API.
}
\examples{
df <- getSeries("D_DUR", country = c("DEU", "FRA"), since = 2008)
head(df)
}

