% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_mapper.R
\name{matrix_mapper}
\alias{matrix_mapper}
\title{Matrix mapper}
\usage{
matrix_mapper(
  xy,
  landscape,
  folder_path,
  moving_window = TRUE,
  window_size,
  restoration_size,
  max_dist,
  habitat = 1,
  matrix = 0,
  plot = FALSE
)
}
\arguments{
\item{xy}{Coordinates of the points (to simulate habitat restoration).}

\item{landscape}{Polygon of the focal landscape.}

\item{folder_path}{Path to the project directory.}

\item{moving_window}{If use or not a moving window over the landscape polygon.
It is very usefull for performance with large data. Default is TRUE.}

\item{window_size}{Size for moving window. Only used if moving_window=TRUE.}

\item{restoration_size}{size of the point in which habitat restoration will be
simulated.}

\item{max_dist}{Parameter for the IIC calculation.}

\item{habitat}{Value in which habitat is assigned in the solution_polygon.}

\item{matrix}{Value in which matrix is assigned in the solution_polygon.}

\item{plot}{If plot or not plot the moving window through the function
iterations. It is usefull for debugging, but consumes computational time and
memmory.}
}
\value{
A new polygon, similiar to the inputed, but with an small point
with additional habitat.
}
\description{
Function that implements an algorithm for maping IIC value at a given (small)
point in the landscape, if it were restored.
}
\examples{
pts = rbind(c(-45,23))
l = rgdal::readOGR('./path to shapefile directory/')
add_points(pts, l, 0.1)

}
