% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchSeries}
\alias{searchSeries}
\title{Search codes and descriptions of available OECD series}
\usage{
searchSeries(data = getDatasets(), string = "unemployment",
  metadata = FALSE)
}
\arguments{
\item{data}{The data frame to search. This can be either a data frame
previously fetched using \code{\link{getDatasets}} (recommended) or left
blank, in which case a temporary data frame is fetched. The second option
adds a few seconds to each search query.}

\item{string}{A string to search for. Can include regular expressions.}

\item{metadata}{Whether or not to include a variable with metadata in the
returned data frame.}
}
\value{
A data frame.
}
\description{
Returns a data frame containing the series codes, descriptions, and
(optionally) metadata for the OECD series which match the given criteria.
}
\examples{
datasets <- getDatasets()
searchSeries(datasets, "unemployment")
}
\seealso{
\code{\link{getDatasets()}}
}

