% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top5.R
\name{Top5}
\alias{Top5}
\title{Top 5 COUNTRY function}
\usage{
Top5(a)
}
\arguments{
\item{a}{This should be a one of column which we want to know top 5.}
}
\value{
Returns a dataframe which includes top 5 countries in OECD and information which we want to know (a)
}
\description{
Top 5 COUNTRY function
}
\examples{
Top5(OECD$AREA)
}
