% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trade.R
\name{Trade}
\alias{Trade}
\title{The energy Trade qty of OECD Countries}
\usage{
Trade(x)
}
\arguments{
\item{x}{"Country name"}
}
\value{
The function shows if this country exports or imports energy and how many energy it exports or imports.
}
\description{
The energy Trade qty of OECD Countries
}
\examples{
Trade("Austrailia")
}
