% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_datasets}
\alias{get_datasets}
\title{Get a data frame with information on all available datasets.}
\usage{
get_datasets(...)
}
\arguments{
\item{...}{Additional parameters passed to \code{data.frame} (e.g.
stringsAsFactors = FALSE).}
}
\value{
A data frame.
}
\description{
Returns a data frame with two variables: \code{id} and \code{description}
}
\examples{
\dontrun{datasets <- get_datasets()}
\dontrun{head(datasets)}

}
\seealso{
\code{\link{search_dataset}} to search for a specific data set or a 
  keyword in the description, and \code{\link{get_data_structure}} to get the
  dimensions of specified data set.
}
