% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDesc}
\alias{getDesc}
\title{Get a list of data frames with variable descriptions.}
\usage{
getDesc(series)
}
\arguments{
\item{series}{A string specifying the code of the series.}
}
\value{
A list of data frames, each corresponding to one variable in the
specified series.
}
\description{
Returns a list of data frames containing descriptions of the variable values
in the requested series.
}
\details{
The data frames returned with \code{\link{getSeries}} often contain variables
whose values are not obvious. For example,
\code{\link{getSeries("MIG_UNEMP_GENDER")}} returns a data frame with the
variable \code{birth} which has the values \code{FB} and \code{NB}, and the
variable \code{rate} which has the values \code{N_RATE} and \code{U_RATE}.
Since the meaning of these are not obvious, \code{getDesc} returns a list of
data frames that can serve as human-readable look-up tables of variables
(see examples below).
}
\examples{
series <- "MIG_UNEM_GENDER"
df <- getSeries(series)
unique(df$birth)
desc <- getDesc(series)
desc$birth
}
\seealso{
\code{\link{getDimensions}} provides similar functionality for
variables names (e.g. getDimensions("MIG_UNEMP_GENDER"), to learn that
the variable \code{birth} refers to "place of birth".
}

