% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{search_dataset}
\alias{search_dataset}
\title{Search codes and descriptions of available OECD series}
\usage{
search_dataset(string, data = get_datasets())
}
\arguments{
\item{string}{A string to search for. Can be a regular expression.}

\item{data}{The data frame to search. This can be either a data frame
previously fetched using \code{\link{get_datasets}} (recommended) or left
blank, in which case a temporary data frame is fetched. The second option
adds a few seconds to each search query.}
}
\value{
A data frame.
}
\description{
Returns a data frame containing the series codes and descriptions for the
OECD series which match the given criteria.
}
\examples{
dsets <- get_datasets()
search_series("employment", dsets)
}
\seealso{
\code{\link{get_datasets}}
}

